chore(cpo): move infrastructure reconciliation to infra package
Move infrastructure-related reconciliation logic from hostedcontrolplane_controller.go
to the infra package for better code organization and testability.

Changes:
- Move reconcileAPIServerService, reconcileOAuthServerService, reconcileKonnectivityServerService,
  reconcileHCPRouterServices, and related functions to infra/infra.go
- Move corresponding tests to infra/infra_test.go
- Add comprehensive test coverage for ReconcileInfrastructure and ReconcileInfrastructureStatus
-- Add fixture-based testing for AWS (Public/Private/PublicAndPrivate) and ARO platforms for Route and LB publishing strategies

Commit-Message-Assisted-by: Claude (via Claude Code)
