+++ control-plane-operator/controllers/hostedcontrolplane/hostedcontrolplane_controller_test.go
@@ -24,7 +24,6 @@ import (
 	"github.com/openshift/hypershift/support/api"
 	fakecapabilities "github.com/openshift/hypershift/support/capabilities/fake"
 	"github.com/openshift/hypershift/support/certs"
-	"github.com/openshift/hypershift/support/config"
 	controlplanecomponent "github.com/openshift/hypershift/support/controlplane-component"
 	"github.com/openshift/hypershift/support/releaseinfo"
 	fakereleaseprovider "github.com/openshift/hypershift/support/releaseinfo/fake"
@@ -37,7 +36,6 @@ import (
 
 	configv1 "github.com/openshift/api/config/v1"
 	"github.com/openshift/api/image/docker10"
-	routev1 "github.com/openshift/api/route/v1"
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/request"
@@ -50,7 +48,6 @@ import (
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime/schema"
-	"k8s.io/apimachinery/pkg/util/intstr"
 	"k8s.io/client-go/util/workqueue"
 	"k8s.io/utils/ptr"
 
@@ -346,552 +343,6 @@ func TestReconcileIgnitionServer(t *testing.T) {
 				if wantKey != gotKey {
 					t.Fatalf("ignition-server-serving-cert data mismatch: want %v, got %v", wantKey, gotKey)
 				}
-			}
-		})
-	}
-}
-
-func TestReconcileOAuthService(t *testing.T) {
-	targetNamespace := "test"
-	apiPort := int32(config.KASSVCPort)
-	hostname := "test.example.com"
-	allowCIDR := []hyperv1.CIDRBlock{"1.2.3.4/24"}
-	ipFamilyPolicy := corev1.IPFamilyPolicyPreferDualStack
-
-	ownerRef := metav1.OwnerReference{
-		APIVersion:         "hypershift.openshift.io/v1beta1",
-		Kind:               "HostedControlPlane",
-		Name:               "test",
-		Controller:         ptr.To(true),
-		BlockOwnerDeletion: ptr.To(true),
-	}
-	oauthPublicService := func(m ...func(*corev1.Service)) corev1.Service {
-		svc := corev1.Service{
-			ObjectMeta: metav1.ObjectMeta{
-				Namespace:       targetNamespace,
-				Name:            manifests.OauthServerService(targetNamespace).Name,
-				OwnerReferences: []metav1.OwnerReference{ownerRef},
-			},
-			Spec: corev1.ServiceSpec{
-				Type:           corev1.ServiceTypeClusterIP,
-				IPFamilyPolicy: &ipFamilyPolicy,
-				Ports: []corev1.ServicePort{
-					{
-						Protocol:   corev1.ProtocolTCP,
-						Port:       apiPort,
-						TargetPort: intstr.FromInt32(apiPort),
-					},
-				},
-				Selector: map[string]string{
-					"app": "oauth-openshift",
-					"hypershift.openshift.io/control-plane-component": "oauth-openshift",
-				},
-			},
-		}
-		for _, m := range m {
-			m(&svc)
-		}
-		return svc
-	}
-	oauthExternalPublicRoute := func(m ...func(*routev1.Route)) routev1.Route {
-		route := routev1.Route{
-			ObjectMeta: metav1.ObjectMeta{
-				Namespace: targetNamespace,
-				Name:      "oauth",
-				Labels: map[string]string{
-					"hypershift.openshift.io/hosted-control-plane": targetNamespace,
-				},
-				OwnerReferences: []metav1.OwnerReference{ownerRef},
-			},
-			Spec: routev1.RouteSpec{
-				Host: hostname,
-				To: routev1.RouteTargetReference{
-					Kind: "Service",
-					Name: manifests.OauthServerService("").Name,
-				},
-				TLS: &routev1.TLSConfig{
-					Termination:                   routev1.TLSTerminationPassthrough,
-					InsecureEdgeTerminationPolicy: routev1.InsecureEdgeTerminationPolicyNone,
-				},
-			},
-		}
-		for _, m := range m {
-			m(&route)
-		}
-		return route
-	}
-	oauthInternalRoute := routev1.Route{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: targetNamespace,
-			Name:      "oauth-internal",
-			Labels: map[string]string{
-				"hypershift.openshift.io/hosted-control-plane": targetNamespace,
-				"hypershift.openshift.io/internal-route":       "true",
-			},
-			OwnerReferences: []metav1.OwnerReference{ownerRef},
-		},
-		Spec: routev1.RouteSpec{
-			Host: "oauth.apps.test.hypershift.local",
-			To: routev1.RouteTargetReference{
-				Kind: "Service",
-				Name: manifests.OauthServerService("").Name,
-			},
-			TLS: &routev1.TLSConfig{
-				Termination:                   routev1.TLSTerminationPassthrough,
-				InsecureEdgeTerminationPolicy: routev1.InsecureEdgeTerminationPolicyNone,
-			},
-		},
-	}
-	testsCases := []struct {
-		name                    string
-		endpointAccess          hyperv1.AWSEndpointAccessType
-		oauthPublishingStrategy hyperv1.ServicePublishingStrategy
-
-		expectedServices []corev1.Service
-		expectedRoutes   []routev1.Route
-	}{
-		{
-			name:           "Route strategy, Public",
-			endpointAccess: hyperv1.Public,
-			oauthPublishingStrategy: hyperv1.ServicePublishingStrategy{
-				Type: hyperv1.Route,
-				Route: &hyperv1.RoutePublishingStrategy{
-					Hostname: hostname,
-				},
-			},
-			expectedServices: []corev1.Service{
-				oauthPublicService(func(s *corev1.Service) {
-					s.Spec.Type = corev1.ServiceTypeClusterIP
-				}),
-			},
-			expectedRoutes: []routev1.Route{
-				oauthExternalPublicRoute(),
-			},
-		},
-		{
-			name:           "Route strategy, PublicPrivate",
-			endpointAccess: hyperv1.PublicAndPrivate,
-			oauthPublishingStrategy: hyperv1.ServicePublishingStrategy{
-				Type: hyperv1.Route,
-				Route: &hyperv1.RoutePublishingStrategy{
-					Hostname: hostname,
-				},
-			},
-
-			expectedServices: []corev1.Service{
-				oauthPublicService(func(s *corev1.Service) {
-					s.Spec.Type = corev1.ServiceTypeClusterIP
-				}),
-			},
-			expectedRoutes: []routev1.Route{
-				oauthExternalPublicRoute(),
-				oauthInternalRoute,
-			},
-		},
-		{
-			name:           "Route strategy, PublicPrivate, no hostname",
-			endpointAccess: hyperv1.PublicAndPrivate,
-			oauthPublishingStrategy: hyperv1.ServicePublishingStrategy{
-				Type: hyperv1.Route,
-			},
-
-			expectedServices: []corev1.Service{
-				oauthPublicService(func(s *corev1.Service) {
-					s.Spec.Type = corev1.ServiceTypeClusterIP
-				}),
-			},
-			expectedRoutes: []routev1.Route{
-				oauthExternalPublicRoute(func(s *routev1.Route) {
-					s.Spec.Host = ""
-					// The route should not be admitted by the private router.
-					delete(s.Labels, "hypershift.openshift.io/hosted-control-plane")
-				}),
-				oauthInternalRoute,
-			},
-		},
-		{
-			name:           "Route strategy, Private",
-			endpointAccess: hyperv1.Private,
-			oauthPublishingStrategy: hyperv1.ServicePublishingStrategy{
-				Type:  hyperv1.Route,
-				Route: &hyperv1.RoutePublishingStrategy{},
-			},
-			expectedServices: []corev1.Service{
-				oauthPublicService(func(s *corev1.Service) {
-					s.Spec.Type = corev1.ServiceTypeClusterIP
-				}),
-			},
-			expectedRoutes: []routev1.Route{
-				oauthInternalRoute,
-			},
-		},
-	}
-	for _, tc := range testsCases {
-		t.Run(tc.name, func(t *testing.T) {
-			hcp := &hyperv1.HostedControlPlane{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: targetNamespace,
-					Name:      "test",
-				},
-				Spec: hyperv1.HostedControlPlaneSpec{
-					Networking: hyperv1.ClusterNetworking{
-						APIServer: &hyperv1.APIServerNetworking{
-							Port:              &apiPort,
-							AllowedCIDRBlocks: allowCIDR,
-						},
-					},
-					Platform: hyperv1.PlatformSpec{
-						Type: hyperv1.AWSPlatform,
-						AWS: &hyperv1.AWSPlatformSpec{
-							EndpointAccess: tc.endpointAccess,
-						},
-					},
-					Services: []hyperv1.ServicePublishingStrategyMapping{{
-						Service:                   hyperv1.OAuthServer,
-						ServicePublishingStrategy: tc.oauthPublishingStrategy,
-					}},
-				},
-			}
-
-			ctx := ctrl.LoggerInto(context.Background(), zapr.NewLogger(zaptest.NewLogger(t)))
-
-			fakeClient := fake.NewClientBuilder().WithScheme(api.Scheme).Build()
-			r := &HostedControlPlaneReconciler{
-				Client: fakeClient,
-				Log:    ctrl.LoggerFrom(ctx),
-			}
-
-			if err := r.reconcileOAuthServerService(ctx, hcp, controllerutil.CreateOrUpdate); err != nil {
-				t.Fatalf("reconcileOAuthServerService failed: %v", err)
-			}
-
-			var actualServices corev1.ServiceList
-			if err := fakeClient.List(ctx, &actualServices); err != nil {
-				t.Fatalf("failed to list services: %v", err)
-			}
-
-			if diff := testutil.MarshalYamlAndDiff(&actualServices, &corev1.ServiceList{Items: tc.expectedServices}, t); diff != "" {
-				t.Errorf("actual services differ from expected: %s", diff)
-			}
-
-			var actualRoutes routev1.RouteList
-			if err := fakeClient.List(ctx, &actualRoutes); err != nil {
-				t.Fatalf("failed to list routes: %v", err)
-			}
-			if diff := testutil.MarshalYamlAndDiff(&actualRoutes, &routev1.RouteList{Items: tc.expectedRoutes}, t); diff != "" {
-				t.Errorf("actual routes differ from expected: %s", diff)
-			}
-		})
-	}
-}
-
-func TestReconcileAPIServerService(t *testing.T) {
-	targetNamespace := "test"
-	apiPort := int32(config.KASSVCPort)
-	kasPort := "client"
-	hostname := "test.example.com"
-	allowCIDR := []hyperv1.CIDRBlock{"1.2.3.4/24"}
-	allowCIDRString := []string{"1.2.3.4/24"}
-	ipFamilyPolicy := corev1.IPFamilyPolicyPreferDualStack
-
-	ownerRef := metav1.OwnerReference{
-		APIVersion:         "hypershift.openshift.io/v1beta1",
-		Kind:               "HostedControlPlane",
-		Name:               "test",
-		Controller:         ptr.To(true),
-		BlockOwnerDeletion: ptr.To(true),
-	}
-	kasPublicService := func(m ...func(*corev1.Service)) corev1.Service {
-		svc := corev1.Service{
-			ObjectMeta: metav1.ObjectMeta{
-				Namespace: targetNamespace,
-				Name:      manifests.KubeAPIServerService(targetNamespace).Name,
-				Annotations: map[string]string{
-					"service.beta.kubernetes.io/aws-load-balancer-type": "nlb",
-					hyperv1.ExternalDNSHostnameAnnotation:               hostname,
-				},
-				Labels: map[string]string{
-					"app": "kube-apiserver",
-					"hypershift.openshift.io/control-plane-component": "kube-apiserver",
-				},
-				OwnerReferences: []metav1.OwnerReference{ownerRef},
-			},
-			Spec: corev1.ServiceSpec{
-				Type:           corev1.ServiceTypeLoadBalancer,
-				IPFamilyPolicy: &ipFamilyPolicy,
-				Ports: []corev1.ServicePort{
-					{
-						Protocol:   corev1.ProtocolTCP,
-						Port:       apiPort,
-						TargetPort: intstr.FromString(kasPort),
-					},
-				},
-				Selector: map[string]string{
-					"app": "kube-apiserver",
-					"hypershift.openshift.io/control-plane-component": "kube-apiserver",
-				},
-			},
-		}
-		for _, m := range m {
-			m(&svc)
-		}
-		return svc
-	}
-	kasPrivateService := func(m ...func(*corev1.Service)) corev1.Service {
-		return kasPublicService(append(m, func(s *corev1.Service) {
-			s.Name = manifests.KubeAPIServerPrivateService(targetNamespace).Name
-
-			delete(s.Annotations, hyperv1.ExternalDNSHostnameAnnotation)
-			s.Annotations["service.beta.kubernetes.io/aws-load-balancer-internal"] = "true"
-
-			s.Labels = nil
-		})...)
-	}
-	withCrossZoneAnnotation := func(svc *corev1.Service) {
-		svc.Annotations["service.beta.kubernetes.io/aws-load-balancer-cross-zone-load-balancing-enabled"] = "true"
-	}
-	withLoadBalancerSourceRanges := func(svc *corev1.Service) {
-		svc.Spec.LoadBalancerSourceRanges = allowCIDRString
-	}
-	kasExternalPublicRoute := routev1.Route{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: targetNamespace,
-			Name:      "kube-apiserver",
-			Labels: map[string]string{
-				"hypershift.openshift.io/hosted-control-plane": targetNamespace,
-			},
-			OwnerReferences: []metav1.OwnerReference{ownerRef},
-		},
-		Spec: routev1.RouteSpec{
-			Host: hostname,
-			To: routev1.RouteTargetReference{
-				Kind: "Service",
-				Name: manifests.KubeAPIServerService("").Name,
-			},
-			TLS: &routev1.TLSConfig{
-				Termination:                   routev1.TLSTerminationPassthrough,
-				InsecureEdgeTerminationPolicy: routev1.InsecureEdgeTerminationPolicyNone,
-			},
-		},
-	}
-	kasExternalPrivateRoute := routev1.Route{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: targetNamespace,
-			Name:      "kube-apiserver-private",
-			Labels: map[string]string{
-				"hypershift.openshift.io/hosted-control-plane": targetNamespace,
-				hyperv1.RouteVisibilityLabel:                   string(hyperv1.RouteVisibilityPrivate),
-				util.InternalRouteLabel:                        "true",
-			},
-			OwnerReferences: []metav1.OwnerReference{ownerRef},
-		},
-		Spec: routev1.RouteSpec{
-			Host: hostname,
-			To: routev1.RouteTargetReference{
-				Kind: "Service",
-				Name: manifests.KubeAPIServerService("").Name,
-			},
-			TLS: &routev1.TLSConfig{
-				Termination:                   routev1.TLSTerminationPassthrough,
-				InsecureEdgeTerminationPolicy: routev1.InsecureEdgeTerminationPolicyNone,
-			},
-		},
-	}
-	kasInternalRoute := routev1.Route{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: targetNamespace,
-			Name:      "kube-apiserver-internal",
-			Labels: map[string]string{
-				"hypershift.openshift.io/hosted-control-plane": targetNamespace,
-				"hypershift.openshift.io/internal-route":       "true",
-			},
-			OwnerReferences: []metav1.OwnerReference{ownerRef},
-		},
-		Spec: routev1.RouteSpec{
-			Host: "api.test.hypershift.local",
-			To: routev1.RouteTargetReference{
-				Kind: "Service",
-				Name: manifests.KubeAPIServerService("").Name,
-			},
-			TLS: &routev1.TLSConfig{
-				Termination:                   routev1.TLSTerminationPassthrough,
-				InsecureEdgeTerminationPolicy: routev1.InsecureEdgeTerminationPolicyNone,
-			},
-		},
-	}
-	testsCases := []struct {
-		name                  string
-		endpointAccess        hyperv1.AWSEndpointAccessType
-		apiPublishingStrategy hyperv1.ServicePublishingStrategy
-
-		expectedServices []corev1.Service
-		expectedRoutes   []routev1.Route
-	}{
-		{
-			name:           "LB strategy, public",
-			endpointAccess: hyperv1.Public,
-			apiPublishingStrategy: hyperv1.ServicePublishingStrategy{
-				Type: hyperv1.LoadBalancer,
-				LoadBalancer: &hyperv1.LoadBalancerPublishingStrategy{
-					Hostname: hostname,
-				},
-			},
-
-			expectedServices: []corev1.Service{
-				kasPublicService(withLoadBalancerSourceRanges),
-			},
-		},
-		{
-			name:           "LB strategy, publicPrivate",
-			endpointAccess: hyperv1.PublicAndPrivate,
-			apiPublishingStrategy: hyperv1.ServicePublishingStrategy{
-				Type: hyperv1.LoadBalancer,
-				LoadBalancer: &hyperv1.LoadBalancerPublishingStrategy{
-					Hostname: hostname,
-				},
-			},
-
-			expectedServices: []corev1.Service{
-				kasPublicService(withCrossZoneAnnotation, withLoadBalancerSourceRanges),
-				kasPrivateService(withCrossZoneAnnotation),
-			},
-		},
-		{
-			name:           "LB strategy, private",
-			endpointAccess: hyperv1.Private,
-			apiPublishingStrategy: hyperv1.ServicePublishingStrategy{
-				Type: hyperv1.LoadBalancer,
-				LoadBalancer: &hyperv1.LoadBalancerPublishingStrategy{
-					Hostname: hostname,
-				},
-			},
-
-			expectedServices: []corev1.Service{
-				kasPublicService(func(s *corev1.Service) {
-					s.Spec.Type = corev1.ServiceTypeClusterIP
-					delete(s.Annotations, "external-dns.alpha.kubernetes.io/hostname")
-				}),
-				kasPrivateService(withCrossZoneAnnotation),
-			},
-		},
-		{
-			name:           "Route strategy, public",
-			endpointAccess: hyperv1.Public,
-			apiPublishingStrategy: hyperv1.ServicePublishingStrategy{
-				Type: hyperv1.Route,
-				Route: &hyperv1.RoutePublishingStrategy{
-					Hostname: hostname,
-				},
-			},
-
-			expectedServices: []corev1.Service{
-				kasPublicService(func(s *corev1.Service) {
-					s.Spec.Type = corev1.ServiceTypeClusterIP
-					delete(s.Annotations, "external-dns.alpha.kubernetes.io/hostname")
-				}),
-			},
-			expectedRoutes: []routev1.Route{
-				kasExternalPublicRoute,
-				kasInternalRoute,
-			},
-		},
-		{
-			name:           "Route strategy, publicPrivate",
-			endpointAccess: hyperv1.PublicAndPrivate,
-			apiPublishingStrategy: hyperv1.ServicePublishingStrategy{
-				Type: hyperv1.Route,
-				Route: &hyperv1.RoutePublishingStrategy{
-					Hostname: hostname,
-				},
-			},
-
-			expectedServices: []corev1.Service{
-				kasPublicService(func(s *corev1.Service) {
-					s.Spec.Type = corev1.ServiceTypeClusterIP
-					delete(s.Annotations, "external-dns.alpha.kubernetes.io/hostname")
-				}),
-			},
-			expectedRoutes: []routev1.Route{
-				kasExternalPublicRoute,
-				kasInternalRoute,
-			},
-		},
-		{
-			name:           "Route strategy, private",
-			endpointAccess: hyperv1.Private,
-			apiPublishingStrategy: hyperv1.ServicePublishingStrategy{
-				Type: hyperv1.Route,
-				Route: &hyperv1.RoutePublishingStrategy{
-					Hostname: hostname,
-				},
-			},
-
-			expectedServices: []corev1.Service{
-				kasPublicService(func(s *corev1.Service) {
-					s.Spec.Type = corev1.ServiceTypeClusterIP
-					delete(s.Annotations, "external-dns.alpha.kubernetes.io/hostname")
-				}),
-			},
-			expectedRoutes: []routev1.Route{
-				kasInternalRoute,
-				kasExternalPrivateRoute,
-			},
-		},
-	}
-	for _, tc := range testsCases {
-		t.Run(tc.name, func(t *testing.T) {
-			hcp := &hyperv1.HostedControlPlane{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: targetNamespace,
-					Name:      "test",
-				},
-				Spec: hyperv1.HostedControlPlaneSpec{
-					Networking: hyperv1.ClusterNetworking{
-						APIServer: &hyperv1.APIServerNetworking{
-							Port:              &apiPort,
-							AllowedCIDRBlocks: allowCIDR,
-						},
-					},
-					Platform: hyperv1.PlatformSpec{
-						Type: hyperv1.AWSPlatform,
-						AWS: &hyperv1.AWSPlatformSpec{
-							EndpointAccess: tc.endpointAccess,
-						},
-					},
-					Services: []hyperv1.ServicePublishingStrategyMapping{{
-						Service:                   hyperv1.APIServer,
-						ServicePublishingStrategy: tc.apiPublishingStrategy,
-					}},
-				},
-			}
-
-			ctx := ctrl.LoggerInto(t.Context(), zapr.NewLogger(zaptest.NewLogger(t)))
-
-			fakeClient := fake.NewClientBuilder().WithScheme(api.Scheme).Build()
-			r := &HostedControlPlaneReconciler{
-				Client: fakeClient,
-				Log:    ctrl.LoggerFrom(ctx),
-			}
-
-			if err := r.reconcileAPIServerService(ctx, hcp, controllerutil.CreateOrUpdate); err != nil {
-				t.Fatalf("reconcileAPIServerService failed: %v", err)
-			}
-
-			var actualServices corev1.ServiceList
-			if err := fakeClient.List(ctx, &actualServices); err != nil {
-				t.Fatalf("failed to list services: %v", err)
-			}
-
-			if diff := testutil.MarshalYamlAndDiff(&actualServices, &corev1.ServiceList{Items: tc.expectedServices}, t); diff != "" {
-				t.Errorf("actual services differ from expected: %s", diff)
-			}
-
-			var actualRoutes routev1.RouteList
-			if err := fakeClient.List(ctx, &actualRoutes); err != nil {
-				t.Fatalf("failed to list routes: %v", err)
-			}
-			if diff := testutil.MarshalYamlAndDiff(&actualRoutes, &routev1.RouteList{Items: tc.expectedRoutes}, t); diff != "" {
-				t.Errorf("actual routes differ from expected: %s", diff)
 			}
 		})
 	}
@@ -1366,142 +817,6 @@ func TestNonReadyInfraTriggersRequeueAfter(t *testing.T) {
 	}
 }
 
-func TestReconcileHCPRouterServices(t *testing.T) {
-	const namespace = "test-ns"
-	publicService := func(m ...func(*corev1.Service)) *corev1.Service {
-		svc := corev1.Service{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "router",
-				Namespace: namespace,
-				Annotations: map[string]string{
-					"service.beta.kubernetes.io/aws-load-balancer-type": "nlb",
-				},
-				Labels: map[string]string{"app": "private-router"},
-			},
-			Spec: corev1.ServiceSpec{
-				Type:     corev1.ServiceTypeLoadBalancer,
-				Selector: map[string]string{"app": "private-router"},
-				Ports: []corev1.ServicePort{
-					{Name: "https", Port: 443, TargetPort: intstr.FromString("https"), Protocol: corev1.ProtocolTCP},
-				},
-			},
-		}
-
-		for _, m := range m {
-			m(&svc)
-		}
-		return &svc
-	}
-	privateService := func(m ...func(*corev1.Service)) *corev1.Service {
-		return publicService(append(m, func(s *corev1.Service) {
-			s.Name = "private-router"
-			s.Annotations["service.beta.kubernetes.io/aws-load-balancer-internal"] = "true"
-		})...)
-	}
-	withCrossZoneAnnotation := func(svc *corev1.Service) {
-		svc.Annotations["service.beta.kubernetes.io/aws-load-balancer-cross-zone-load-balancing-enabled"] = "true"
-	}
-	tests := []struct {
-		name                         string
-		endpointAccess               hyperv1.AWSEndpointAccessType
-		exposeAPIServerThroughRouter bool
-		existingObjects              []client.Object
-		expectedServices             []corev1.Service
-	}{
-		{
-			name:                         "Public HCP gets public LB only",
-			endpointAccess:               hyperv1.Public,
-			exposeAPIServerThroughRouter: true,
-			expectedServices: []corev1.Service{
-				*publicService(),
-			},
-		},
-		{
-			name:                         "PublicPrivate gets public and private LB",
-			endpointAccess:               hyperv1.PublicAndPrivate,
-			exposeAPIServerThroughRouter: true,
-			expectedServices: []corev1.Service{
-				*privateService(withCrossZoneAnnotation),
-				*publicService(withCrossZoneAnnotation),
-			},
-		},
-		{
-			name:                         "Private gets private LB only",
-			endpointAccess:               hyperv1.Private,
-			exposeAPIServerThroughRouter: true,
-			expectedServices: []corev1.Service{
-				*privateService(withCrossZoneAnnotation),
-			},
-		},
-		{
-			name:                         "Public LB gets removed when switching to Private",
-			endpointAccess:               hyperv1.Private,
-			exposeAPIServerThroughRouter: true,
-			existingObjects:              []client.Object{publicService(), privateService()},
-			expectedServices: []corev1.Service{
-				*privateService(withCrossZoneAnnotation),
-			},
-		},
-		{
-			name:                         "No LB created when public and not using Route",
-			endpointAccess:               hyperv1.Public,
-			exposeAPIServerThroughRouter: false,
-			expectedServices:             nil,
-		},
-	}
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			hcp := &hyperv1.HostedControlPlane{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "hcp",
-					Namespace: namespace,
-				},
-				Spec: hyperv1.HostedControlPlaneSpec{
-					Platform: hyperv1.PlatformSpec{
-						Type: hyperv1.AWSPlatform,
-						AWS: &hyperv1.AWSPlatformSpec{
-							EndpointAccess: tc.endpointAccess,
-						},
-					},
-				},
-			}
-			if tc.exposeAPIServerThroughRouter {
-				hcp.Spec.Services = []hyperv1.ServicePublishingStrategyMapping{
-					{
-						Service: hyperv1.APIServer,
-						ServicePublishingStrategy: hyperv1.ServicePublishingStrategy{
-							Type: hyperv1.Route,
-							Route: &hyperv1.RoutePublishingStrategy{
-								Hostname: "apiserver.example.com",
-							},
-						},
-					},
-				}
-			}
-
-			ctx := ctrl.LoggerInto(t.Context(), zapr.NewLogger(zaptest.NewLogger(t)))
-			c := fake.NewClientBuilder().WithScheme(api.Scheme).WithObjects(append(tc.existingObjects, hcp)...).Build()
-
-			r := HostedControlPlaneReconciler{
-				Client: c,
-				Log:    ctrl.LoggerFrom(ctx),
-			}
-
-			if err := r.reconcileHCPRouterServices(ctx, hcp, controllerutil.CreateOrUpdate); err != nil {
-				t.Fatalf("reconcileRouter failed: %v", err)
-			}
-
-			var services corev1.ServiceList
-			if err := c.List(ctx, &services); err != nil {
-				t.Fatalf("failed to list services: %v", err)
-			}
-			if diff := testutil.MarshalYamlAndDiff(&services, &corev1.ServiceList{Items: tc.expectedServices}, t); diff != "" {
-				t.Errorf("actual services differ from expected: %s", diff)
-			}
-		})
-	}
-}
-
 func TestSetKASCustomKubeconfigStatus(t *testing.T) {
 	hcp := sampleHCP(t)
 	pullSecret := &corev1.Secret{ObjectMeta: metav1.ObjectMeta{Namespace: hcp.Namespace, Name: "pull-secret"}}
@@ -1663,106 +978,6 @@ func TestIncludeServingCertificates(t *testing.T) {
 			} else {
 				g.Expect(err).NotTo(HaveOccurred())
 				g.Expect(newRootCA.Data["ca.crt"]).To(Equal(tc.expectedCert))
-			}
-		})
-	}
-}
-
-type fakeMessageCollector struct {
-	msg string
-}
-
-func (c *fakeMessageCollector) ErrorMessages(resource client.Object) ([]string, error) {
-	return []string{c.msg}, nil
-}
-
-func TestReconcileRouterServiceStatus(t *testing.T) {
-	const namespace = "test-ns"
-	const svcName = "test"
-	tests := []struct {
-		name         string
-		svc          *corev1.Service
-		expectedHost string
-		expectMsg    bool
-	}{
-		{
-			name: "Non-existent service",
-		},
-		{
-			name: "Service that has not been provisioned",
-			svc: &corev1.Service{
-				ObjectMeta: metav1.ObjectMeta{Name: svcName, Namespace: namespace},
-			},
-			expectMsg: true,
-		},
-		{
-			name: "Service with host populated",
-			svc: &corev1.Service{
-				ObjectMeta: metav1.ObjectMeta{Name: svcName, Namespace: namespace},
-				Status: corev1.ServiceStatus{
-					LoadBalancer: corev1.LoadBalancerStatus{
-						Ingress: []corev1.LoadBalancerIngress{
-							{
-								Hostname: "test.host",
-							},
-						},
-					},
-				},
-			},
-			expectedHost: "test.host",
-		},
-		{
-			name: "Service with IP populated",
-			svc: &corev1.Service{
-				ObjectMeta: metav1.ObjectMeta{Name: svcName, Namespace: namespace},
-				Status: corev1.ServiceStatus{
-					LoadBalancer: corev1.LoadBalancerStatus{
-						Ingress: []corev1.LoadBalancerIngress{
-							{
-								IP: "1.2.3.4",
-							},
-						},
-					},
-				},
-			},
-			expectedHost: "1.2.3.4",
-		},
-	}
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			ctx := ctrl.LoggerInto(t.Context(), zapr.NewLogger(zaptest.NewLogger(t)))
-			existing := []client.Object{}
-			if tc.svc != nil {
-				existing = append(existing, tc.svc)
-			}
-			c := fake.NewClientBuilder().WithScheme(api.Scheme).WithObjects(existing...).Build()
-
-			r := HostedControlPlaneReconciler{
-				Client: c,
-				Log:    ctrl.LoggerFrom(ctx),
-			}
-			msgCollector := &fakeMessageCollector{msg: "test message"}
-			svc := &corev1.Service{
-				ObjectMeta: metav1.ObjectMeta{Name: svcName, Namespace: namespace},
-			}
-			host, needed, msg, err := r.reconcileRouterServiceStatus(ctx, svc, msgCollector)
-			if err != nil {
-				t.Fatalf("unexpected err: %v", err)
-			}
-			if !needed {
-				t.Fatalf("unexpected, needed == false")
-			}
-			if host != tc.expectedHost {
-				t.Errorf("unexpected host, actual: %s, expected: %s", host, tc.expectedHost)
-			}
-			if tc.expectMsg {
-				if msg == "" {
-					t.Errorf("did not get an event message")
-				}
-			} else {
-				if len(msg) > 0 {
-					t.Errorf("got unexpected event message")
-				}
 			}
 		})
 	}
